workflow/requirements.txt
alembic==0.9.10
apache-airflow==1.10.3
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
Babel==2.7.0
bandit==1.6.0
black==19.3b0
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
configparser==3.5.3
coverage==4.5.3
croniter==0.3.30
defusedxml==0.6.0
dill==0.2.9
docutils==0.14
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
Flask-Admin==1.5.3
Flask-AppBuilder==1.12.3
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.11.1
google-auth==1.6.3
google-cloud-core==1.0.1
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.17
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.10
json-merge-patch==0.2
lockfile==0.12.2
lxml==4.3.3
Mako==1.0.11
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
mypy==0.701
mypy-extensions==0.4.1
numpy==1.16.4
ordereddict==1.1
packaging==19.0
pandas==0.24.2
pbr==5.2.1
pendulum==1.4.4
pluggy==0.12.0
protobuf==3.8.0
psutil==5.6.2
psycopg2-binary==2.8.2
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.4.2
pyparsing==2.4.0
pytest==4.6.2
pytest-cov==2.7.1
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.1
PyYAML==5.1
requests==2.22.0
rope==0.14.0
rsa==4.0
selenium==3.141.0
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.2.19
stevedore==1.30.1
tabulate==0.8.3
tenacity==4.12.0
text-unidecode==1.2
thrift==0.11.0
toml==0.10.0
typed-ast==1.3.5
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
zipp==0.5.1
zope.deprecation==4.4.0
